% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_similarity.R
\name{compare_similarity}
\alias{compare_similarity}
\title{Compare Data Distributions}
\usage{
compare_similarity(real_data, synthetic_data)
}
\arguments{
\item{real_data}{Original dataset.}

\item{synthetic_data}{Synthetic dataset.}
}
\value{
Tibble with means by variable for real and synthetic data and the corresponding KS p-value.
}
\description{
Compares distributions of real and synthetic data using a Kolmogorov-Smirnov test.
}
\examples{
structure <- describe_structure(mtcars)
synth <- generate_synthetic_sample(n = 2, structure = structure, prompt = "Simulate realistic car performance data with similar distributions.")
compare_similarity(mtcars, synth)
}
