% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_structure.R
\name{describe_structure}
\alias{describe_structure}
\title{Describe data structure}
\usage{
describe_structure(data, n_examples = 100)
}
\arguments{
\item{data}{One or more data frames of real respondent profiles.}

\item{n}{Number of examples to show from each column.}
}
\value{
A summary of each variable.
}
\description{
Summarizes the structure of a data frame, recording variable name, data types, and n sample values.
}
\examples{
structure <- describe_structure(mtcars, n = 10)
}
