% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_response.R
\name{gen_response}
\alias{gen_response}
\title{Generate a Simulated Interview Response Based on Profile Inputted}
\usage{
gen_response(
  txt_file,
  respondent_profile,
  question,
  k_similar = 10,
  model = "gpt-4o",
  embedding_model = "text-embedding-3-small",
  base_url = "https://api.openai.com/v1",
  api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{txt_file}{Text file containing interview segments. Each interview should be separated by a line containing only "---".}

\item{respondent_profile}{A brief description of the respondent's profile to guide the simulated response.}

\item{question}{The prompt or question to simulate responses for.}

\item{k_similar}{Number of most similar respondents to use as reference when generating. Default is 10.}

\item{model}{The model to use for the chat (defaults to "gpt-4o").}

\item{embedding_model}{The embedding model to use (defaults to "text-embedding-3-small").}

\item{base_url}{The base URL to the endpoint. Default is "https://api.openai.com/v1".}

\item{api_key}{API key for the embedding and chat service. Default reads from \code{Sys.getenv("OPENAI_API_KEY")}.}
}
\value{
A simulated response from the respondent based on the most relevant interview segments.
}
\description{
This function loads interview segments from a text file, embeds each segment,
finds the most relevant segments to a given question, and then uses the OPENAI chat model to produce
a realistic response as if from a specific respondent.
}
\examples{
\dontrun{
response <- gen_response(
txt_file = "cces_observations.txt",
respondent_profile = "30-year-old Asian-American female.",
k_similar = 3,
question = "What do you know about the Luka Doncic trade to the Lakers?")
}

}
