% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_synthetic_dataset.R
\name{gen_synthetic_dataset}
\alias{gen_synthetic_dataset}
\title{Generate a synthetic dataset}
\usage{
gen_synthetic_dataset(
  n = 10,
  structure = NULL,
  prompt = NULL,
  model = "GPT-4.1",
  base_url = "https://litellmproxy.osu-ai.org/"
)
}
\arguments{
\item{n}{Number of rows of synthetic data to generate.}

\item{structure}{The output from \code{describe_structure} function telling the generator what columns and datatypes to use.}

\item{prompt}{Instructions to give the LLM}

\item{model}{The model to use for the chat (defaults to "gpt-4.1")}

\item{base_url}{The base URL to the endpoint. Default is "https://litellmproxy.osu-ai.org".}
}
\value{
A data frame containing synthetic values, with the same variables (columns) as the original data.
}
\description{
Creates a dataset with synthetic data that matches the variables, distributions, ranges, and data types of the input structure.
}
\examples{
structure <- describe_structure(mtcars)
synth <- gen_synthetic_sample(n = 2, structure = structure, prompt = "Simulate realistic car performance data with similar distributions.")
}
