% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_synthetic_panel.R
\name{gen_synthetic_panel}
\alias{gen_synthetic_panel}
\title{Generate a Synthetic Panel of Respondents}
\usage{
gen_synthetic_panel(
  ...,
  n = 50,
  question =
    "Simulate a survey question about a generic preference, rating, or behavior.",
  model = "GPT-4.1",
  base_url = "https://litellmproxy.osu-ai.org/"
)
}
\arguments{
\item{...}{One or more data frames of real respondent profiles to base the synthetic panel on.}

\item{n}{Number of rows to generate. Default is 50.}

\item{question}{The prompt or question the synthetic respondents will answer.}

\item{model}{The model to use for the chat (defaults to "gpt-4.1")}

\item{base_url}{The base URL to the endpoint. Default is "https://litellmproxy.osu-ai.org".}

\item{variations}{String vector describing a different variation.}
}
\value{
A synthetic panel of n respondents with simulated responses to a survey question.
}
\description{
Creates new fictional respondents based on input profiles and simulates their responses to a survey question.
The synthetic respondents match the distributions and ranges of the provided profiles.
}
\examples{
test1 <- data.frame(
  age = c(23, 45, 38, 52, 33, 33),
  gender = c("F", "M", "M", "F", "M", "M"),
  income = c(44000, 72000, 1000, 84000, 35000, 111000))

test2 <- data.frame(
  educ = c("BS", "MD", "HS", "PhD", "BA"),
  party = c("R", "R", "D", "I", "D"))

synth_panel <- gen_synthetic_panel(
  test1, test2, n = 10,
  question = "How do you feel about calling over texting?")
}
